requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-pubsub
google-cloud-storage==1.42.2
gcsfs==2021.9.0
google-cloud-datastore==2.7.1
Flask==2.1.2